// src/components/molecules/dialog/components/dialog-container/enums/dialog-container.component.enums.ts
enum DialogContainerComponentBorderRadiusEnum {
  LARGE = 'large',
  MEDIUM = 'medium',
  NONE = 'none',
  SMALL = 'small',
}

enum DialogContainerComponentPaddingEnum {
  LARGE = 'large',
  MEDIUM = 'medium',
  NONE = 'none',
  SMALL = 'small',
}

export { DialogContainerComponentBorderRadiusEnum, DialogContainerComponentPaddingEnum };